# encoding/json: fix and optimize marshal for quoted string