@breml 2019-09-06
@mvdan Thanks for your hint about the benchmarks. From my point of view, the change was not primary driven by performance optimization in mind, this came more as a welcome side effect.

My motivation for the proposed change is:

1. remove the unnecessary error case
2. fix #34154

Nevertheless I will try to have a look at `benchstat` and provide some benchmark statistics for the improved performance as well.