@breml 2019-09-06
@mvdan I had a look at the benchmarks in `encoding/json` and non of them contains struct fields with the string option set in the struct tags.

The change affects the encoding of struct like:

```Go
type Foo struct {
	Bar string `json:"bar,string"`
}
```

While looking through the commit messages for the recent changes on `encoding/json` I found this [one](https://github.com/golang/go/commit/ae68a912725e5a3a0482bc5945687663f2ddafe3#diff-cbb71488de8e250b59cee0de7af128f6) from you:

> ...
> This change should slightly reduce the amount of work when decoding into
> json.Number, though that isn't very common nor part of any current
> benchmarks.
> ...

I would argue, that this change also slightly improves encoding performance and that it is not very common nor part of any existing benchmarks. Would this be fine as well or do you insist on benchmarks for this particular case?