@mvdan 2019-09-06
Please reply on Gerrit. I'd rather not keep the discussion in two places.